requirements.txt
async-generator==1.10
attrs==22.1.0
awscli==1.25.71
boto3==1.24.70
botocore==1.27.70
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
docutils==0.16
Flask==2.2.2
Flask-Cors==3.0.10
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.23.3
outcome==1.2.0
pandas==1.4.4
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
